dipex/requirements.txt
Django==1.6.4
django-jsonfield==0.9.13
